Create medical_education
There are several benefits of studying MBBS in the Philippines. The living cost of this country is relatively low compared to India, with market prices being about 1.75% higher. While transportation charges are low, electronic products are three times more expensive. However, clothing and food prices are generally comparable. The Philippines is a safe choice for students from India who wish to gain experience in the medical field. Nevertheless, it's important to consider the cost of medical insurance when considering whether to study in the country.